created 2 separate table rows with river data

+++ db/migrate/20160829205607_station.rb
@@ -2,7 +2,7 @@ class Station < ActiveRecord::Migration[5.0]
   def change
     create_table :stations do |t|
       t.string :station_name
-      t.integer :gage_height
+      t.decimal :gage_height
       t.integer :discharge
       t.integer :river_id
     end
